// frontend/src/app/core/api/model/teacher-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TeacherRequest { 
    firstName: string;
    lastName: string;
    firstNameAr?: string;
    lastNameAr?: string;
    email?: string;
    phone?: string;
    cin?: string;
    gender?: TeacherRequest.GenderEnum;
    contractType: TeacherRequest.ContractTypeEnum;
    hireDate: string;
    degree: TeacherRequest.DegreeEnum;
    specialtyIds: Set<number>;
}
export namespace TeacherRequest {
    export const GenderEnum = {
        Male: 'MALE',
        Female: 'FEMALE'
    } as const;
    export type GenderEnum = typeof GenderEnum[keyof typeof GenderEnum];
    export const ContractTypeEnum = {
        Cdi: 'CDI',
        Cdd: 'CDD',
        Freelance: 'FREELANCE',
        Guest: 'GUEST'
    } as const;
    export type ContractTypeEnum = typeof ContractTypeEnum[keyof typeof ContractTypeEnum];
    export const DegreeEnum = {
        Bachelor: 'BACHELOR',
        FourYearBachelor: 'FOUR_YEAR_BACHELOR',
        Master: 'MASTER',
        EngineeringDegree: 'ENGINEERING_DEGREE',
        Doctorate: 'DOCTORATE'
    } as const;
    export type DegreeEnum = typeof DegreeEnum[keyof typeof DegreeEnum];
}